Gave choices and pages their own flag sets
Before, we were saving the page and choices arrays as part of the save file in order to preserve which page/choice had been disabled.

This approached had a fatal flaw: it would overwrite the pages and choices in memory. Normally, this would not be a problem, however, if you added any pages, everything would go bonkers. This is still not iron clad as adding new pages might cause the flag ids to not match anymore but at least it won't be as catastrophic as before.