Created Speaker editor
Created a new "speaker" editor so that feeds into the dialogue editor.  This makes it so we don't have to use names and image ids when defining dialogues. It's the first step before updating the dialogue editor to allow for each page to have a different speaker.

Also created a base class for the editor windows, there is a lot of copy/paste boilerplate I need to start removing.

Still need to do three things before this phase is complete:
- Need to update locations to say which speakers can be found there (so we don't load all the speakers, right now it would be 100kb if chip ram which is toooo much)
- Need to make the changes in the engine to actually be able to use this information.
- Remove the old shapes from the locations