Redid how speakers are handled
Completely changed how speakers in the dialogues are handled.
Created a speaker editor to define speakers instead of using the shape editor
Update the location editor to allow you to define which speakers are there
Created a wrapper file format (MPSH) around the Blitz shapes, this will make it possible to do stuff that LoadShapes simply can't

One step closer to redoing dialogues!