Cutscene check as new music player
Added a new checkbox in the scene editor that let's you define if this scene is part of a cutscene. If so, the PDA icon won't be rendered, but more crucially, the game won't crash if you click it (or where it should be) during a cutscene.

Changed the music library to be more compatible with AROS

Added a new title screen, though it's not hooked up yet

Updated Italian strings